fix(*): align inference kwargs + dtype to canonical HF model-card recipe
User's PT path mode-collapsed on <ref>\"\"\"\"\" even after lm_head was correctly
tied. Pulled the exact inference snippet from
https://huggingface.co/nvidia/LocateAnything-3B and diffed against ours.

Canonical (verbatim from model card LocateAnythingWorker.predict):
    dtype = torch.bfloat16
    response = self.model.generate(
        ...
        max_new_tokens=2048,
        do_sample=True,
        temperature=0.7,
        top_p=0.9,
        repetition_penalty=1.1,
    )

Ours (before this commit):
    REF_DTYPE = torch.float16
    out = self.model.generate(
        ...
        do_sample=False,
        repetition_penalty=1.1,
    )

Two deviations, both individually known to cause mode-collapse on Qwen2-family
generators:

  1) fp16 dtype: Qwen2's RoPE / attention-score path can overflow in fp16
     (max ~65504), producing NaN/Inf hidden states that surface as
     <ref>$$$$$ or <ref>\"\"\"\"\" at decode time. bfloat16 has fp32's exponent
     range and avoids the overflow. This is a well-documented issue on
     Llama/Qwen-family models.

  2) do_sample=False: greedy decoding on MTP/hybrid generation traps the
     model in a fixed-point token (the per-step argmax). repetition_penalty=
     1.1 only penalizes EXACT repeats of the just-emitted token, not the
     broader degeneration pattern. Temperature sampling at T=0.7 breaks the
     fixed point.

Changes:
  - config.py: REF_DTYPE = torch.float16 → torch.bfloat16
  - orchestrator._detect_via_pt: do_sample=True, temperature=0.7, top_p=0.9
  - orchestrator._detect_via_trt: temperature=0.0,top_p=1.0 → 0.7, 0.9
  - pipelines.run_compare PT branch: same canonical kwargs

(The TRT pixel_values path explicitly casts to np.float16 for the fixed-
resolution TRT engine input — that cast is independent of REF_DTYPE and
unchanged.)

Tests: 113 passed (no behavioral test changes needed — kwargs are runtime).

Why this wasn't found earlier:
  - I kept guessing root causes instead of re-reading the model card
  - User had to remind me twice that the PT weights are identical to the paper
  - The shim suppressed the loudest symptom (lm_head AttributeError) which
    pulled investigation in a different direction

+++ lrai_locate_anything/config.py
@@ -14,7 +14,11 @@ from typing import Optional
 import torch
 
 MODEL_ID = "nvidia/LocateAnything-3B"
-REF_DTYPE = torch.float16
+# bfloat16 per the canonical inference recipe on the HF model card. fp16 silently
+# overflows in Qwen2's RoPE / attention-score path on this model (max ~65504),
+# producing NaN/Inf in hidden states and <ref>$$$$$ mode-collapse at decode time.
+# bfloat16 has fp32's exponent range so there is no overflow.
+REF_DTYPE = torch.bfloat16
 
 # MoonViT structural constants (do not change unless the upstream model does)
 ENG_PATCH_SIZE = 14
